+++ packages/apollo-server-core/src/errors.ts
@@ -199,6 +199,17 @@ export class ForbiddenError extends ApolloError {
   }
 }
 
+export class BadUserInputError extends ApolloError {
+  constructor(message: string, properties?: Record<string, any>) {
+    super(message, 'BAD_USER_INPUT', properties);
+
+    // Set the prototype explicitly.
+    // https://stackoverflow.com/a/41102306
+    Object.setPrototypeOf(this, BadUserInputError.prototype);
+    Object.defineProperty(this, 'name', { value: 'BadUserInputError' });
+  }
+}
+
 export function formatApolloErrors(
   errors: Array<Error>,
   options?: {